Update visitor code
We need to account for the fact that multiple items can lead to the same node type, so code from two distinct items will actually go in the same visitor function.
To accommodate this, an item number has been added as a field in the base node type, and the reduction functions set this when constructing the node.
The visitor functions are then assembled by putting a switch statement in the generated visitor code, the cases of which are the specified code from
the parser description.